oscar/cdk/requirements.txt
# OSCAR CDK Requirements
# This file contains all dependencies with security constraints applied

# Core CDK dependencies
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# AWS SDK with explicit version control to prevent vulnerable urllib3
boto3==1.39.13
botocore<1.40.0,>=1.39.13

# Security fixes - explicitly pin secure urllib3 version
# Addresses CVE-2025-50181 and CVE-2025-50182 (urllib3 1.26.20 vulnerabilities)
urllib3==2.5.0

# Additional security constraints for transitive dependencies
requests>=2.31.0
certifi>=2023.7.22